+++ api/lib/auth.ts
@@ -1,9 +1,10 @@
 import { Request } from 'express';
+import { TSchema } from '@sinclair/typebox';
 import Err from '@openaddresses/batch-error';
 import jwt from 'jsonwebtoken';
 import Config from './config.js';
 import { InferSelectModel } from 'drizzle-orm';
-import { Profile } from './schema.js';
+import { Profile, Connection } from './schema.js';
 
 export enum ResourceCreationScope {
     SERVER = 'server',
@@ -12,6 +13,7 @@ export enum ResourceCreationScope {
 
 export enum AuthUserAccess {
     ADMIN = 'admin',
+    AGENCY = 'agency',
     USER = 'user'
 }
 
@@ -77,10 +79,13 @@ export default class Auth {
     /**
      * Is the requester authenticated - can be either a Resource or User auth
      *
-     * @param {Object} req Express Request
-     * @param {boolean} token Should URL query tokens be allowed (usually only for downloads)
+     * @param config    - Server Config
+     * @param req       - Express Request
+     * @param opts.token        - Should URL query tokens be allowed (usually only for downloads)
+     * @param opts.anyResources - Any Resource token can use this endpoint
+     * @param resources         - Array of resource types that can use this endpoint
      */
-    static async is_auth(config: Config, req: Request<unknown, unknown, unknown, any>, opts: {
+    static async is_auth(config: Config, req: Request<any, any, any, any>, opts: {
         token?: boolean;
         anyResources?: boolean;
         resources?: Array<AuthResourceAccepted>;
@@ -99,7 +104,7 @@ export default class Auth {
             const auth_resource = auth as AuthResource;
 
             if (opts.anyResources && opts.resources.length) {
-                throw new Err(403, null, 'Server cannot specify defined resource access any resource access together');
+                throw new Err(403, null, 'Server cannot specify defined resource access and any resource access together');
             } else if (!opts.anyResources && !opts.resources.length) {
                 throw new Err(403, null, 'Resource token cannot access resource');
             }
@@ -108,7 +113,11 @@ export default class Auth {
                 try {
                     await config.models.ConnectionToken.from(auth_resource.token);
                 } catch (err) {
-                    throw new Err(403, err.name === 'PublicError' ? err : new Error(String(err)), 'Token does not exist');
+                    if (err instanceof Error) {
+                        throw new Err(403, err.name === 'PublicError' ? err : new Error(String(err)), 'Token does not exist');
+                    } else {
+                        throw new Err(500, new Error(String(err)), 'Unknown Token Error');
+                    }
                 }
             }
 
@@ -122,7 +131,38 @@ export default class Auth {
         return auth;
     }
 
-    static async is_user(config: Config, req: Request<unknown, unknown, unknown, any>): Promise<boolean> {
+    static async is_connection(config: Config, req: Request<any, any, any, any>, opts: {
+        token?: boolean;
+        resources?: Array<AuthResourceAccepted>;
+    }, connectionid: number): Promise<{
+        auth: AuthResource | AuthUser;
+        connection: InferSelectModel<typeof Connection>;
+        profile?: InferSelectModel<typeof Profile>;
+    }> {
+        const auth = await this.is_auth(config, req, opts)
+
+        const connection = await config.models.Connection.from(connectionid);
+
+        if (this.#is_user(auth)) {
+            const profile = await this.#as_profile(config, auth as AuthUser);
+
+            if (profile.system_admin === true) {
+                return { connection, profile, auth };
+            } else {
+                if (!connection.agency) throw new Err(401, null, 'Only a System Admin can access this connection');
+                if (!profile.agency_admin) throw new Err(401, null, 'Only an Agency Admin admin or higher can access connections');
+                if (!profile.agency_admin.includes(connection.agency)) throw new Err(401, null, `You are not an Agency Admin for Agency ${connection.agency}`);
+
+                return { connection, profile, auth };
+            }
+        } else {
+            // If a resource token is used it's up to the caller to specify it is allowed via the resources array
+            // is_auth will disallow resource tokens when no resource array is set
+            return { auth, connection };
+        }
+    }
+
+    static async is_user(config: Config, req: Request<any, any, any, any>): Promise<boolean> {
         const auth = await this.is_auth(config, req);
         return this.#is_user(auth);
     }
@@ -131,7 +171,7 @@ export default class Auth {
         return auth instanceof AuthUser;
     }
 
-    static async is_resource(config: Config, req: Request<unknown, unknown, unknown, any>): Promise<boolean> {
+    static async is_resource(config: Config, req: Request<any, any, any, any>): Promise<boolean> {
         const auth = await this.is_auth(config, req);
         return this.#is_resource(auth);
     }
@@ -140,7 +180,9 @@ export default class Auth {
         return auth instanceof AuthResource;
     }
 
-    static async as_resource(config: Config, req: Request<unknown, unknown, unknown, any>, opts: {
+    static async as_resource(config: Config, req: Request<any, any, any, any>, opts: {
+        anyResources?: boolean;
+        resources?: Array<AuthResourceAccepted>;
         token?: boolean;
     } = {}): Promise<AuthResource> {
         if (!opts.token) opts.token = false;
@@ -151,7 +193,7 @@ export default class Auth {
         return auth as AuthResource;
     }
 
-    static async as_user(config: Config, req: Request<unknown, unknown, unknown, any>, opts: {
+    static async as_user(config: Config, req: Request<any, any, any, any>, opts: {
         token?: boolean;
         admin?: boolean;
     } = {}): Promise<AuthUser> {
@@ -169,16 +211,20 @@ export default class Auth {
         return user;
     }
 
-    static async as_profile(config: Config, req: Request<unknown, unknown, unknown, any>, opts: {
+    static async #as_profile(config: Config, user: AuthUser): Promise<InferSelectModel<typeof Profile>> {
+        return await config.models.Profile.from(user.email);
+    }
+
+    static async as_profile(config: Config, req: Request<any, any, any, any>, opts: {
         token?: boolean;
         admin?: boolean;
     } = {}): Promise<InferSelectModel<typeof Profile>> {
         const user = await this.as_user(config, req, opts);
-        return await config.models.Profile.from(user.email);
+        return await this.#as_profile(config, user);
     }
 }
 
-function auth_request(config: Config, req: Request<unknown, unknown, unknown, any>, opts?: {
+function auth_request(config: Config, req: Request<any, any, any, any>, opts?: {
     token: boolean
 }): AuthResource | AuthUser {
     try {
@@ -218,10 +264,15 @@ function auth_request(config: Config, req: Request<unknown, unknown, unknown, an
                     throw err;
                 }
             }
+        } else {
+            throw new Err(401, null, 'No Auth Present')
         }
     } catch (err) {
-        if (err.name === 'PublicError') throw err;
-        throw new Err(401, err instanceof Error ? err : new Error(String(err)), 'Invalid Token')
+        if (err instanceof Error && err.name === 'PublicError') {
+            throw err;
+        } else {
+            throw new Err(401, new Error(String(err)), 'Invalid Token')
+        }
     }
 }