+++ api/web/src/stores/map.ts
@@ -501,7 +501,9 @@ function cotStyles(id: string, opts: {
             'line-width': ["*", 2, ["number", ["get", "stroke-width"], 3]],
             'line-opacity': ["number", ["get", "stroke-opacity"], 1]
         }
-    },{
+    }])
+
+    const circle = {
         id: id,
         type: 'circle',
         source: id,
@@ -512,7 +514,13 @@ function cotStyles(id: string, opts: {
             'circle-radius': ["number", ["get", "circle-radius"], 4],
             'circle-opacity': ["number", ["get", "circle-opacity"], 1]
         }
-    }])
+    }
+
+    if (opts.group) {
+        circle.filter.push(['!has', 'group']);
+    }
+
+    styles.push(circle);
 
     if (opts.icons) {
         styles.push({